Employee CRUD

+++ WebApplication/Models/Employee.cs
@@ -12,7 +12,7 @@ namespace WebApplication.Models
             InverseReportsToNavigation = new HashSet<Employee>();
         }
 
-        public short EmployeeId { get; set; }
+        public int EmployeeId { get; set; }
         public string LastName { get; set; }
         public string FirstName { get; set; }
         public string Title { get; set; }
@@ -28,7 +28,7 @@ namespace WebApplication.Models
         public string Extension { get; set; }
         public byte[] Photo { get; set; }
         public string Notes { get; set; }
-        public short? ReportsTo { get; set; }
+        public int? ReportsTo { get; set; }
         public string PhotoPath { get; set; }
 
         public virtual Employee ReportsToNavigation { get; set; }
